+++ Block/Adminhtml/Form/Edit/Tab/NewsletterSettings.php
@@ -1,0 +1,93 @@
+<?php
+/**
+ * Copyright © Alekseon sp. z o.o.
+ * http://www.alekseon.com/
+ */
+namespace Alekseon\WidgetForms\Block\Adminhtml\Form\Edit\Tab;
+
+use Alekseon\AlekseonEav\Model\Adminhtml\System\Config\Source\InputType;
+use Alekseon\AlekseonEav\Api\Data\EntityInterface;
+use Magento\Framework\Data\Form\Element\AbstractElement;
+
+/**
+ * Class NewsletterSettings
+ * @package Alekseon\WidgetForms\Block\Adminhtml\Form\Edit\Tab
+ */
+class NewsletterSettings extends \Alekseon\AlekseonEav\Block\Adminhtml\Entity\Edit\Form implements
+    \Magento\Backend\Block\Widget\Tab\TabInterface
+{
+    protected $dataObject;
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getTabLabel()
+    {
+        return __('Newsletter Settings');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getTabTitle()
+    {
+        return __('Newsletter Settings');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function canShowTab()
+    {
+        return (bool) $this->getDataObject()->getCanUseForWidget();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function isHidden()
+    {
+        return false;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getDataObject()
+    {
+        if (null === $this->dataObject) {
+            return $this->_coreRegistry->registry('current_form');
+        }
+        return $this->dataObject;
+    }
+
+    /**
+     * @return $this
+     * @throws \Magento\Framework\Exception\LocalizedException
+     */
+    protected function _prepareForm()
+    {
+        $dataObject = $this->getDataObject();
+
+        /** @var \Magento\Framework\Data\Form $form */
+        $form = $this->_formFactory->create();
+        $newsletterFieldset = $form->addFieldset('newsletter_settings_fieldset', ['legend' => __('Newletter Settings')]);
+        $this->addAllAttributeFields($newsletterFieldset, $dataObject,['included' => ['newsletter']]);
+
+        $this->setForm($form);
+
+        return parent::_prepareForm();
+    }
+
+
+    /**
+     * Initialize form fileds values
+     *
+     * @return $this
+     */
+    protected function _initFormValues()
+    {
+        $this->getForm()->addValues($this->getDataObject()->getData());
+        return parent::_initFormValues();
+    }
+}