+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - checkbox intup block type for multiselect
+- subscribe to newsetter
 ### Changed
 - Removed widget forms from menu. Now its done by CustomFormsBuilder module